--- rawData.csv
@@ -1,26 +1,0 @@
-4.728,3.843,3.869,3.874,3.864,3.878,3.85,3.852,3.882,3.902,3.871,3.82,3.872,3.858,3.799,3.858,3.868,3.852,3.779,3.844,3.85,3.846,3.876,3.884,3.851,3.852
-3.686,3.634,3.66,3.686,3.648,3.719,3.696,3.776,3.709,3.715,3.683,3.705,3.704,3.734,3.699,3.72,3.739,3.706,3.737,3.687,3.708,3.731,3.725,3.703,3.75,3.674
-3.683,3.625,3.606,3.672,3.636,3.664,3.689,3.699,3.7,3.67,3.677,3.662,3.704,3.747,3.757,3.706,3.739,3.707,3.706,3.743,3.714,3.729,3.683,3.698,3.71,3.741
-3.711,3.631,3.632,3.669,3.674,3.634,3.65,3.655,3.674,3.656,3.648,3.668,3.625,3.721,3.699,3.699,3.692,3.727,3.734,3.717,3.73,3.719,3.744,3.712,3.753,3.722
-3.678,3.628,3.63,3.623,3.621,3.633,3.626,3.673,3.684,3.627,3.676,3.624,3.66,3.704,3.699,3.662,3.693,3.696,3.695,3.701,3.673,3.752,3.751,3.727,3.716,3.726
-3.656,3.688,3.596,3.626,3.628,3.628,3.607,3.669,3.672,3.666,3.613,3.665,3.629,3.635,3.648,3.651,3.665,3.654,3.685,3.669,3.69,3.683,3.694,3.721,3.695,3.696
-3.697,3.659,3.619,3.657,3.633,3.617,3.653,3.643,3.629,3.628,3.611,3.665,3.666,3.635,3.646,3.628,3.666,3.652,3.658,3.654,3.681,3.669,3.672,3.687,3.715,3.663
-3.725,3.634,3.648,3.611,3.584,3.625,3.636,3.629,3.619,3.648,3.637,3.658,3.621,3.65,3.644,3.658,3.663,3.681,3.637,3.661,3.67,3.655,3.673,3.647,3.668,3.653
-3.703,3.641,3.625,3.59,3.659,3.577,3.615,3.598,3.627,3.595,3.694,3.63,3.647,3.662,3.67,3.638,3.702,3.624,3.647,3.633,3.644,3.647,3.692,3.642,3.648,3.668
-3.709,3.664,3.66,3.627,3.623,3.681,3.62,3.627,3.563,3.624,3.633,3.668,3.651,3.656,3.64,3.635,3.641,3.663,3.685,3.616,3.696,3.659,3.685,3.664,3.662,3.679
-3.674,3.75,3.622,3.626,3.643,3.67,3.619,3.651,3.604,3.608,3.642,3.651,3.629,3.643,3.622,3.646,3.613,3.66,3.66,3.654,3.679,3.627,3.658,3.67,3.65,3.639
-3.692,3.701,3.625,3.638,3.605,3.63,3.651,3.592,3.656,3.633,3.616,3.675,3.605,3.589,3.64,3.656,3.656,3.634,3.619,3.678,3.66,3.631,3.653,3.666,3.647,3.663
-3.695,3.737,3.661,3.636,3.604,3.629,3.625,3.635,3.603,3.63,3.632,3.654,3.617,3.63,3.652,3.648,3.654,3.661,3.682,3.631,3.643,3.654,3.692,3.643,3.652,3.652
-3.681,3.727,3.64,3.633,3.617,3.623,3.61,3.629,3.578,3.63,3.602,3.641,3.639,3.651,3.603,3.641,3.65,3.605,3.627,3.657,3.636,3.658,3.657,3.63,3.643,3.666
-3.658,3.688,3.619,3.665,3.612,3.633,3.608,3.576,3.617,3.603,3.652,3.647,3.626,3.628,3.624,3.641,3.63,3.638,3.658,3.62,3.673,3.666,3.63,3.67,3.675,3.647
-3.702,3.761,3.669,3.656,3.643,3.618,3.617,3.627,3.688,3.605,3.588,3.604,3.608,3.64,3.615,3.642,3.623,3.639,3.652,3.591,3.646,3.629,3.68,3.664,3.689,3.635
-3.67,3.729,3.645,3.611,3.651,3.621,3.652,3.586,3.653,3.654,3.605,3.626,3.588,3.637,3.636,3.608,3.642,3.589,3.643,3.654,3.602,3.63,3.62,3.614,3.663,3.658
-3.692,3.731,3.674,3.62,3.637,3.63,3.619,3.643,3.658,3.6,3.592,3.593,3.602,3.611,3.621,3.675,3.643,3.652,3.615,3.626,3.595,3.622,3.637,3.647,3.628,3.647
-3.677,3.717,3.693,3.646,3.635,3.634,3.632,3.667,3.637,3.643,3.623,3.591,3.651,3.616,3.634,3.626,3.641,3.626,3.613,3.661,3.628,3.629,3.639,3.606,3.672,3.668
-3.675,3.751,3.771,3.674,3.63,3.649,3.649,3.595,3.642,3.643,3.628,3.602,3.622,3.616,3.621,3.623,3.637,3.613,3.639,3.609,3.651,3.631,3.653,3.641,3.648,3.623
-3.663,3.745,3.708,3.633,3.605,3.677,3.647,3.65,3.638,3.659,3.628,3.622,3.608,3.657,3.642,3.617,3.613,3.615,3.655,3.632,3.635,3.597,3.631,3.643,3.68,3.638
-3.678,3.705,3.749,3.645,3.612,3.677,3.652,3.615,3.64,3.646,3.642,3.593,3.649,3.619,3.613,3.65,3.626,3.67,3.641,3.608,3.622,3.609,3.607,3.609,3.626,3.682
-3.703,3.726,3.721,3.604,3.636,3.631,3.632,3.597,3.642,3.632,3.62,3.634,3.618,3.617,3.615,3.634,3.642,3.63,3.629,3.623,3.632,3.667,3.613,3.621,3.617,3.604
-3.673,3.734,3.718,3.618,3.631,3.661,3.603,3.621,3.619,3.609,3.63,3.679,3.625,3.654,3.606,3.607,3.614,3.601,3.604,3.65,3.636,3.62,3.626,3.598,3.704,3.638
-3.679,3.747,3.711,3.668,3.682,3.625,3.652,3.626,3.651,3.604,3.604,3.614,3.664,3.633,3.643,3.636,3.6,3.601,3.602,3.612,3.616,3.616,3.65,3.634,3.602,3.627
-3.706,3.73,3.796,3.671,3.644,3.635,3.656,3.638,3.632,3.633,3.594,3.678,3.618,3.631,3.625,3.599,3.61,3.628,3.641,3.637,3.601,3.641,3.621,3.626,3.657,3.615
